use runMode method in runMode.js inside browser and node
