+++ addon/runmode/runmode-node.js
@@ -237,7 +237,12 @@ StringStream.prototype.baseToken = function () {
   return oracle && oracle.baseToken(this.pos)
 };
 
-var runmodeNode = CodeMirror = Object.assign({}, modesMethods, {StringStream: StringStream});
+// Create a minimal CodeMirror needed to use runMode.
+CodeMirror = Object.assign({}, modesMethods, {StringStream: StringStream});
+
+// Shim library CodeMirror with the minimal CodeMirror defined above. 
+require.cache[require.resolve("../../lib/codemirror")] = require.cache[require.resolve("./runmode.node")];
+require.cache[require.resolve("../../addon/runmode/runmode")] = require.cache[require.resolve("./runmode.node")];  
 
 // Minimal default mode.
 CodeMirror.defineMode("null", function () { return ({token: function (stream) { return stream.skipToEnd(); }}); });
@@ -246,15 +251,72 @@ CodeMirror.defineMIME("text/plain", "null");
 CodeMirror.registerHelper = CodeMirror.registerGlobalHelper = Math.min;
 CodeMirror.splitLines = function(string) { return string.split(/\r?\n|\r/); };
 
+CodeMirror.defaults = { indentUnit: 2 };
+
+// CodeMirror, copyright (c) by Marijn Haverbeke and others
+// Distributed under an MIT license: https://codemirror.net/LICENSE
+
+(function(mod) {
+  if (typeof exports == "object" && typeof module == "object") // CommonJS
+    { mod(require("../../lib/codemirror")); }
+  else if (typeof define == "function" && define.amd) // AMD
+    { define(["../../lib/codemirror"], mod); }
+  else // Plain browser env
+    { mod(CodeMirror); }
+})(function(CodeMirror) {
+
 CodeMirror.runMode = function(string, modespec, callback, options) {
-  var mode = CodeMirror.getMode({ indentUnit: 2 }, modespec);
-  var tabSize = (options && options.tabSize) || 4;
+  var mode = CodeMirror.getMode(CodeMirror.defaults, modespec);
+  var ie = /MSIE \d/.test(navigator.userAgent);
+  var ie_lt9 = ie && (document.documentMode == null || document.documentMode < 9);
+
+  if (callback.appendChild) {
+    var tabSize = (options && options.tabSize) || CodeMirror.defaults.tabSize;
+    var node = callback, col = 0;
+    node.innerHTML = "";
+    callback = function(text, style) {
+      if (text == "\n") {
+        // Emitting LF or CRLF on IE8 or earlier results in an incorrect display.
+        // Emitting a carriage return makes everything ok.
+        node.appendChild(document.createTextNode(ie_lt9 ? '\r' : text));
+        col = 0;
+        return;
+      }
+      var content = "";
+      // replace tabs
+      for (var pos = 0;;) {
+        var idx = text.indexOf("\t", pos);
+        if (idx == -1) {
+          content += text.slice(pos);
+          col += text.length - pos;
+          break;
+        } else {
+          col += idx - pos;
+          content += text.slice(pos, idx);
+          var size = tabSize - col % tabSize;
+          col += size;
+          for (var i = 0; i < size; ++i) { content += " "; }
+          pos = idx + 1;
+        }
+      }
+
+      if (style) {
+        var sp = node.appendChild(document.createElement("span"));
+        sp.className = "cm-" + style.replace(/ +/g, " cm-");
+        sp.appendChild(document.createTextNode(content));
+      } else {
+        node.appendChild(document.createTextNode(content));
+      }
+    };
+  }
 
   var lines = CodeMirror.splitLines(string), state = (options && options.state) || CodeMirror.startState(mode);
-  var oracle = {lookAhead: function(n) { return lines[i + n] }, baseToken: function() {}};
   for (var i = 0, e = lines.length; i < e; ++i) {
     if (i) { callback("\n"); }
-    var stream = new CodeMirror.StringStream(lines[i], tabSize, oracle);
+    var stream = new CodeMirror.StringStream(lines[i], null, {
+      lookAhead: function(n) { return lines[i + n] },
+      baseToken: function() {}
+    });
     if (!stream.string && mode.blankLine) { mode.blankLine(state); }
     while (!stream.eol()) {
       var style = mode.token(stream, state);
@@ -264,7 +326,4 @@ CodeMirror.runMode = function(string, modespec, callback, options) {
   }
 };
 
-require.cache[require.resolve("../../lib/codemirror")] = require.cache[require.resolve("./runmode.node")];
-require.cache[require.resolve("../../addon/runmode/runmode")] = require.cache[require.resolve("./runmode.node")];
-
-module.exports = runmodeNode;
+});
