+++ package.json
@@ -25,7 +25,8 @@
     "node-static": "0.7.11",
     "puppeteer": "^1.20.0",
     "rollup": "^1.26.3",
-    "@rollup/plugin-buble": "^0.21.3"
+    "@rollup/plugin-buble": "^0.21.3",
+    "@rollup/plugin-multi-entry": "^3.0.1"
   },
   "bugs": "http://github.com/codemirror/CodeMirror/issues",
   "keywords": [
@@ -43,5 +44,6 @@
     "dependencies": {},
     "devDependencies": {}
   },
-  "dependencies": {}
+  "dependencies": {
+  }
 }
